Make App a context manager
This means it can be used either as it is now unchanged or like this:

with libui.App():
    ... # code

Note that (1) the build instructions for libui appear to be wrong "make" vs "cmake ."; and (2) I can't build libui because of a bug in it or Ubuntu 14.04's cmake I don't know which. So this change is untested.